% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.sentence.df.R
\name{as.sentence.df}
\alias{as.sentence.df}
\title{A function for turning a body-data.frame into a sentence-data.frame}
\usage{
as.sentence.df(df, col = "body")
}
\arguments{
\item{col}{Which column contains the sentences?}
}
\description{
This function turns a data.frame containing multiple-sentence text, into a 
data frame containing the sentences on separate rows, while keeping metadata.
}
\examples{
as.sentence.df() 
}
\keyword{NLP}
