% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzkeep.R
\name{fuzzScore}
\alias{fuzzScore}
\title{Fuzzily Score Sentences}
\usage{
fuzzScore(df, col, lookup, verbose = FALSE)
}
\arguments{
\item{df}{Data frame to apply to}

\item{col}{Column in data frame to apply to}

\item{lookup}{A character vector of strings to match against}

\item{verbose}{print timing}
}
\description{
Scores each string in df[col] by calculating string distance between it and each string in 
lookup. Adds the best match and score to data frame.
}
\examples{
mtcars$nstring <- names(mtcars)
check <- c('Porch','Auto','Hello','World')
fuzzScore(mtcars,'nstring',check) 
}
\keyword{Text}
\keyword{processing}
