% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadTextRows.R
\name{spreadTextRows}
\alias{spreadTextRows}
\title{Spread Text Rows}
\usage{
spreadTextRows(df, col = "body", FUN, ...)
}
\arguments{
\item{col}{Column to apply to}

\item{FUN}{Function to apply}

\item{...}{Arguments to FUN}
}
\description{
This function applies FUN to each text cell in col, which returns a list of results.
This list is then spread into separate rows, each retaining the original values in the
other rows. Originally used to split sentences using a tokenizer.
}
\examples{
SpreadTextRows(iris$species,FUN) 
}
\keyword{Text}
\keyword{processing}
